Move HeadingLevel to tree.
