+++ ftml/src/tree/mod.rs
@@ -21,12 +21,14 @@
 mod anchor;
 mod container;
 mod element;
+mod heading;
 mod list;
 mod module;
 
 pub use self::anchor::*;
 pub use self::container::*;
 pub use self::element::*;
+pub use self::heading::*;
 pub use self::list::*;
 pub use self::module::*;
